created dir for Progress components moved Progress.js into Progress dir modified imports accordingly
