+++ client/src/index.js
@@ -15,7 +15,7 @@ import PasswordReset from "./components/PasswordReset";
 import LandingPage from "./components/LandingPage";
 import Schedule from "./components/Schedule";
 import Workouts from "./components/Workouts";
-import Progress from "./components/Progress";
+import Progress from "./components/Progress/Progress";
 import Billing from "./components/Billing";
 import Settings from "./components/Settings";
